Add printfs to identify bug
There is something wrong with get_contour function and Temika. Still
trying to get to the bottom of it. I have also copied the function that
works from last-stable.

I tried subbing it in to see if that fixes things, but it does not! I am
led to think that perhaps there's something amiss with the ct_st
structure.